+++ src/main/java/com/tristankechlo/toolleveling/ToolLevelingClient.java
@@ -3,6 +3,7 @@ package com.tristankechlo.toolleveling;
 import com.tristankechlo.toolleveling.client.blockentityrenderer.ToolLevelingTableRenderer;
 import com.tristankechlo.toolleveling.client.screen.ToolLevelingTableHandledScreen;
 import com.tristankechlo.toolleveling.network.ClientNetworkHandler;
+import com.tristankechlo.toolleveling.utils.Names;
 import com.tristankechlo.toolleveling.utils.Names.NetworkChannels;
 
 import net.fabricmc.api.ClientModInitializer;
@@ -17,7 +18,7 @@ import org.apache.logging.log4j.Logger;
 @Environment(EnvType.CLIENT)
 public final class ToolLevelingClient implements ClientModInitializer {
 
-	public static final Logger LOGGER = LogManager.getLogger();
+	public static final Logger LOGGER = LogManager.getLogger(Names.MOD_NAME);
 
 	@Override
 	public void onInitializeClient() {